Cache method results
We then annotate our Service beans to cache
their results. The important aspect here is that
we need to configure the EhCache cache with
the same name that is used by `@Cacheable`
here. So in ehcache.xml we also have a cache
configuration with name `bars`.

Notice that for this specific example, the cache
has a short time to live (10 seconds) so that we
can see it being refreshed.